Add licence headers, move Encryption classes to same package
