+++ sync-android/src/main/java/com/cloudant/sync/sqlite/android/encryption/EncryptedPreparedAttachment.java
@@ -1,3 +1,17 @@
+/**
+ * Copyright (c) 2015 IBM Cloudant, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ * either express or implied. See the License for the specific language governing permissions
+ * and limitations under the License.
+ */
+
 package com.cloudant.sync.sqlite.android.encryption;
 
 import com.cloudant.sync.datastore.Attachment;
@@ -6,8 +20,6 @@ import com.cloudant.sync.datastore.AttachmentNotSavedException;
 import com.cloudant.sync.util.Misc;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
 import java.util.UUID;
 
 /**
